Add e2e test for validator ordering
