+++ .circleci/config.yml
@@ -598,6 +598,24 @@ jobs:
             cd packages/celotool
             ./ci_test_attestations.sh checkout master
 
+  end-to-end-geth-validator-order-test:
+    <<: *e2e-defaults
+    resource_class: medium+
+    steps:
+      - attach_workspace:
+          at: ~/app
+      - run:
+          name: Check if the test should run
+          command: |
+            FILES_TO_CHECK="${PWD}/packages/celotool,${PWD}/packages/protocol,${PWD}/.circleci/config.yml"
+            ./scripts/ci_check_if_test_should_run_v2.sh ${FILES_TO_CHECK}
+      - run:
+          name: Run test
+          command: |
+            set -e
+            cd packages/celotool
+            ./ci_test_validator_order.sh checkout master
+
   web:
     working_directory: ~/app
     docker:
@@ -727,6 +745,10 @@ workflows:
             - lint-checks
             - contractkit-test
       - end-to-end-geth-attestations-test:
+          requires:
+            - lint-checks
+            - contractkit-test
+      - end-to-end-geth-validator-order-test:
           requires:
             - lint-checks
             - contractkit-test
